+++ CHANGELOG.md
@@ -5,6 +5,7 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 ## [Unreleased]
 ### Changed
 - Remove deprecated methods in Client.
+- Do not default to primary marketplace.
 - Update list of error codes created up front.
 - Do not support Ruby 2.2.